wip: cleans up github code
